+++ src/World/CharacterList.h
@@ -25,7 +25,7 @@
 class CharacterList
 {
     public:
-        void Populate(uint8 count, WorldPacket &recvPacket);
+        void Populate(uint32 count, WorldPacket &recvPacket);
         void Print() const;
  
         Character const* GetCharacterByName(std::string name) const;
